refactor(lambda): move zip file logic to its own function and file
