+++ packages/artillery/lib/platform/aws-lambda/index.js
@@ -2,11 +2,6 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-const temp = require('temp');
-const fs = require('fs-extra');
-const spawn = require('cross-spawn');
-const chalk = require('chalk');
-
 const EventEmitter = require('events');
 const debug = require('debug')('platform:aws-lambda');
 
@@ -14,18 +9,13 @@ const { randomUUID } = require('crypto');
 
 const sleep = require('../../util/sleep');
 const path = require('path');
-
-const archiver = require('archiver');
 const AWS = require('aws-sdk');
 
 const https = require('https');
 
 const { QueueConsumer } = require('../../queue-consumer');
 
-const { createBOM } = require('../../create-bom/create-bom');
-
 const setDefaultAWSCredentials = require('../aws/aws-set-default-credentials');
-const { promisify } = require('node:util');
 
 const telemetry = require('../../telemetry').init();
 const crypto = require('node:crypto');
@@ -38,6 +28,7 @@ const ensureS3BucketExists = require('../aws/aws-ensure-s3-bucket-exists');
 const getAccountId = require('../aws/aws-get-account-id');
 
 const createSQSQueue = require('../aws/aws-create-sqs-queue');
+const { createAndUploadLambdaZip } = require('./dependencies');
 
 // https://stackoverflow.com/a/66523153
 function memoryToVCPU(memMB) {
@@ -121,21 +112,6 @@ class PlatformLambda {
     await setDefaultAWSCredentials(AWS);
     this.accountId = await getAccountId();
 
-    const dirname = temp.mkdirSync(); // TODO: May want a way to override this by the user
-    const zipfile = temp.path({ suffix: '.zip' });
-
-    debug({ dirname, zipfile });
-
-    let createBomOpts = {};
-    let entryPoint = this.opts.absoluteScriptPath;
-    let extraFiles = [];
-    if (this.opts.absoluteConfigPath) {
-      entryPoint = this.opts.absoluteConfigPath;
-      extraFiles.push(this.opts.absoluteScriptPath);
-      createBomOpts.entryPointIsConfig = true;
-    }
-    // TODO: custom package.json path here
-
     const metadata = {
       region: this.region,
       platformConfig: {
@@ -145,39 +121,19 @@ class PlatformLambda {
     };
     global.artillery.globalEvents.emit('metadata', metadata);
 
-    artillery.log('- Bundling test data');
-    const bom = await promisify(createBOM)(
-      entryPoint,
-      extraFiles,
-      createBomOpts
+    //make sure the bucket exists to send the zip file or the dependencies to
+    const bucketName = await ensureS3BucketExists(
+      this.region,
+      this.s3LifecycleConfigurationRules
     );
-    for (const f of bom.files) {
-      artillery.log('  -', f.noPrefix);
-    }
-
-    // Copy handler:
-    fs.copyFileSync(
-      path.resolve(__dirname, 'lambda-handler', 'index.js'),
-      path.join(dirname, 'index.js')
+    this.bucketName = bucketName;
+    const { bom, s3path } = await createAndUploadLambdaZip(
+      this.bucketName,
+      this.opts.absoluteScriptPath,
+      this.opts.absoluteConfigPath,
+      this.platformOpts.cliArgs.dotenv
     );
-    fs.copyFileSync(
-      path.resolve(__dirname, 'lambda-handler', 'package.json'),
-      path.join(dirname, 'package.json')
-    );
-
-    // FIXME: This may overwrite lambda-handler's index.js or package.json
-    // Copy files that make up the test:
-    for (const o of bom.files) {
-      fs.ensureFileSync(path.join(dirname, o.noPrefix));
-      fs.copyFileSync(o.orig, path.join(dirname, o.noPrefix));
-    }
-
-    if (this.platformOpts.cliArgs.dotenv) {
-      fs.copyFileSync(
-        path.resolve(process.cwd(), this.platformOpts.cliArgs.dotenv),
-        path.join(dirname, path.basename(this.platformOpts.cliArgs.dotenv))
-      );
-    }
+    this.lambdaZipPath = s3path;
 
     this.artilleryArgs.push('run');
 
@@ -227,118 +183,6 @@ class PlatformLambda {
       (x) => x.orig === this.opts.absoluteScriptPath
     )[0];
     this.artilleryArgs.push(p.noPrefix);
-
-    artillery.log('- Installing dependencies');
-    const { stdout, stderr, status, error } = spawn.sync(
-      'npm',
-      ['install', '--omit', 'dev'],
-      {
-        cwd: dirname
-      }
-    );
-
-    if (error) {
-      artillery.log(stdout?.toString(), stderr?.toString(), status, error);
-    } else {
-      // artillery.log('        npm log is in:', temp.path({suffix: '.log'}));
-    }
-
-    // Install extra plugins & engines
-    if (bom.modules.length > 0) {
-      artillery.log(
-        `- Installing extra engines & plugins: ${bom.modules.join(', ')}`
-      );
-      const { stdout, stderr, status, error } = spawn.sync(
-        'npm',
-        ['install'].concat(bom.modules),
-        { cwd: dirname }
-      );
-      if (error) {
-        artillery.log(stdout?.toString(), stderr?.toString(), status, error);
-      }
-    }
-
-    // Copy this version of Artillery into the Lambda package
-    const a9basepath = path.resolve(__dirname, '..', '..', '..');
-    // TODO: read this from .files in package.json instead:
-    for (const dir of ['bin', 'lib']) {
-      const destdir = path.join(dirname, 'node_modules', 'artillery', dir);
-      const srcdir = path.join(a9basepath, dir);
-      fs.ensureDirSync(destdir);
-      fs.copySync(srcdir, destdir);
-    }
-    for (const fn of ['console-reporter.js', 'util.js']) {
-      const destfn = path.join(dirname, 'node_modules', 'artillery', fn);
-      const srcfn = path.join(a9basepath, fn);
-      fs.copyFileSync(srcfn, destfn);
-    }
-
-    fs.copyFileSync(
-      path.resolve(a9basepath, 'package.json'),
-      path.join(dirname, 'node_modules', 'artillery', 'package.json')
-    );
-
-    const a9cwd = path.join(dirname, 'node_modules', 'artillery');
-    debug({ a9basepath, a9cwd });
-
-    const {
-      stdout: stdout2,
-      stderr: stderr2,
-      status: status2,
-      error: error2
-    } = spawn.sync('npm', ['install', '--omit', 'dev'], { cwd: a9cwd });
-    if (error2) {
-      artillery.log(stdout2?.toString(), stderr2?.toString(), status2, error2);
-    } else {
-      // artillery.log('        npm log is in:', temp.path({suffix: '.log'}));
-    }
-
-    const {
-      stdout: stdout3,
-      stderr: stderr3,
-      status: status3,
-      error: error3
-    } = spawn.sync(
-      'npm',
-      [
-        'uninstall',
-        'dependency-tree',
-        'detective',
-        'is-builtin-module',
-        'try-require',
-        'walk-sync',
-        'esbuild-wasm',
-        'artillery-plugin-publish-metrics'
-      ],
-      {
-        cwd: a9cwd
-      }
-    );
-    if (error3) {
-      artillery.log(stdout3?.toString(), stderr3?.toString(), status3, error3);
-    } else {
-      // artillery.log('        npm log is in:', temp.path({suffix: '.log'}));
-    }
-
-    fs.removeSync(path.join(dirname, 'node_modules', 'aws-sdk'));
-    fs.removeSync(path.join(a9cwd, 'node_modules', 'typescript'));
-    fs.removeSync(path.join(a9cwd, 'node_modules', 'tap'));
-    fs.removeSync(path.join(a9cwd, 'node_modules', 'prettier'));
-
-    artillery.log('- Creating zip package');
-    await this.createZip(dirname, zipfile);
-
-    artillery.log('Preparing AWS environment...');
-    const bucketName = await ensureS3BucketExists(
-      this.region,
-      this.s3LifecycleConfigurationRules
-    );
-    this.bucketName = bucketName;
-
-    const s3path = await this.uploadLambdaZip(bucketName, zipfile);
-    debug({ s3path });
-    this.lambdaZipPath = s3path;
-
     // 36 is length of a UUUI v4 string
     const queueName = `${SQS_QUEUES_NAME_PREFIX}_${this.testRunId.slice(
       0,
@@ -643,21 +487,6 @@ class PlatformLambda {
     }
   }
 
-  async createZip(src, out) {
-    const archive = archiver('zip', { zlib: { level: 9 } });
-    const stream = fs.createWriteStream(out);
-
-    return new Promise((resolve, reject) => {
-      archive
-        .directory(src, false)
-        .on('error', (err) => reject(err))
-        .pipe(stream);
-
-      stream.on('close', () => resolve());
-      archive.finalize();
-    });
-  }
-
   async createLambdaRole() {
     const ROLE_NAME = 'artilleryio-default-lambda-role-20230116';
     const POLICY_NAME = 'artilleryio-lambda-policy-20230116';
@@ -776,21 +605,6 @@ class PlatformLambda {
     }
 
     await lambda.createFunction(lambdaConfig).promise();
-  }
-
-  async uploadLambdaZip(bucketName, zipfile) {
-    const key = `lambda/${randomUUID()}.zip`;
-    // TODO: Set lifecycle policy on the bucket/key prefix to delete after 24 hours
-    const s3 = new AWS.S3();
-    const s3res = await s3
-      .putObject({
-        Body: fs.createReadStream(zipfile),
-        Bucket: bucketName,
-        Key: key
-      })
-      .promise();
-
-    return key;
   }
 }
